Backend/requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-classic
langgraph
langsmith
openai
pinecone
numpy
tiktoken
PyPDF2
python-dotenv
boto3
fastapi
uvicorn[standard]
python-multipart
pydantic
ddgs
beautifulsoup4
requests
websockets>=10.0